pyproject.toml
[project]
name = "brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "litellm>=1.83.14",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.12",
]
